tee-time-bot-root-2/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
httpx==0.28.1
beautifulsoup4==4.12.3
aiosqlite==0.20.0
apscheduler==3.10.4
python-telegram-bot==21.9
cryptography==44.0.0
python-dotenv==1.0.1
playwright==1.49.1
pydantic==2.10.4
